improvement(testsuite): add "test_misc_py" unit test
Add "test_misc.py" to have a place for unit tests.

Python's unit test frameworks are powerful and simple. We miss out by not
adding plenty of tests for our plain python code.

So far, there is only a trivial test to see whether it runs. The idea is
to extend that in the future, as we add more tests.

The name "misc" is chosen intentionally/arbitrarily. I don't think there
is value in overthinking the place where a individual unit tests reside.
We just add a test, somewhere, where usually it just passes and runs
forever. We don't need to care, unless a test breaks. And when it
breaks, we can see very clearly where this happens, without being
confused that everything is just in "test_misc.py" -- the large heap of
unrelated tests.