improvement(testsuite): automatically pick source files for tests
Previously, the existing tests just include the firewalld module, and
expect it to find it in the PYTHONPATH=. There is even a comment, that
if the user want to use the sources from the git tree, to explicitly set
this.

This should just happen automatically.

During `import fwtst`, append the source directory to `sys.path` so that
we import firewalld modules from the source tree.

Note that the tests also get installed and packaged separately. So
`fwtst` has a (simple) heuristic to detect when it runs from the source
tree or from an installation. If that heuristic gets it wrong, the user
can still overwrite it with `FWTST_AUTO_PYTHONPATH=0` and set
`PYTHONPATH=` explicitly.

Obviously, `import fwtst` must happen before importing other firewalld
modules, as this sets up the `PYTHONPATH=`