improvement(testsuite): add pytest dependency for unit tests
I think we should significantly expand our unit tests. There are two
popular unit test frameworks in Python, "unittest" and "pytest". The
former is part of the standard library, which is a benefit.

Pytest is very popular and often considered to be the "best" unit test
framework for python (source: internet). I think we should use it.

Adding new dependencies is something that we should aim to avoid or do
carefully. But the pytest dependency is only for running the unit tests
and/or a build dependency. That makes it more acceptable.