improvement(policy): return policy objects from get_policies()
Return Policy objects instead of just the name. We can easily get the name
from the policy object, if we need it. And often we actually need the
object itself. Returning the name seems more limited.

In terms of performance, it might make little difference. Calling
p_obj.name is probably also a dictionary lookup, just like a subsequent
get_policy() is (as of recently). Still, it makes more sense that functions
return the more useful data (the entire policy object), when they already go
through the effort to fetch it.