improvement(policy): return policy object from get_policies_not_derived_from_zone()
Return Policy objects instead of just the name. We can easily get the name
from the policy object, if we need it. And often we actually need the
object itself. Returning the name seems more limited.

In terms of performance, it might make little difference. Calling
p_obj.name is probably also a dictionary lookup, just like get_policy()
is (as of recently). Still, it makes more sense that functions return
the more useful data (the entire policy object), when they already go
through the effort to fetch it.

Also, sorted() takes an iterable and returns a new list. This seems
wasteful, we already constructed a new list, so we could just have
called policies.sort() to sort the list inplace. Instead, use a
generator expression and sorted().