improvement(zone): add Ident class for hashing the ID of zone/policy
We index zones and policies by their plain names.

For example, there are the special zones named "ANY" and "HOST".
However, the user is able to create a zone with such names. As we only
compare the zone's name, there is a conflict.

Also, for zones we create policys via policy_obj_from_zone_obj(),
with names "zone_{fromZone}_{toZone}" (where either fromZone or toZone
is one of "ANY"/"HOST"). But nothing prevents the user to create a
policy with such name. Well, if the zone "public" already exists,
and so does "zone_public_ANY", then there will be a failure to add a
policy "zone_public_ANY". However, if the user first adds a policy
"zone_foo_ANY" and then adds a zone "foo", bad things happen.

Also, the user can crate a zone "ANY". Which results in the creation of
two policies "zone_ANY_ANY" (which of course does not work, as the
policy name is the identifier for the policy). Again, bad things happen.

Add an immutable Ident type, which can contain identifiers for those
uses.

Will be used next.