improvement(glibutil): add firewall.glibutil.timeout helper for keeping track of timers
When we schedule a timer, we must always keep track of it so that it can
be cancelled or restarted.

firewalld is centered around the GLib mainloop via `dbus.mainloop.glib`.
The new API uses GLib timers and GMainContext for the timeouts. It is
thus placed in a new, GLib-dependent module "firewall.glibutil".

That GLib timer API returns a handle (a number) with which we can cancel
the timeout later. The number is slightly inconvenient, because we must
never cancel a timer that was already cancelled or invoked. That could
be avoided by using GSource instances instead of the numeric source IDs.
Still, it means that the caller needs to keep track of the handle.

The new API allows to instead pass key and tags, to cancel timers
without need to keeping track of the handle. The key uniquely identifies
the timer, and a caller might know in advance a unique key they can use,
instead of tracking the handle. Additionally, tags allow to bulk cancel
timers. These features should make it simpler to manage pending timers,
and allow to forget about the handle. Keeping track of the handle means
to keep track of another piece of state, instead let the "timeout" API
keep track of it.

The use case is that we schedule timeouts for a zone's service. So we
can

   timeout.schedule(timeout, callback, key=('service', zone_name, service_name))

using a caller-chosen key that we can later also use for cancellation. We can also
pass tags, to bulk cancel timers when the zone/service gets deleted, via

   timeout.schedule(timeout,
                    callback,
                    key=('service', zone_name, service_name),
                    tags=(('zone', zone_name), ('service', service_name)))